Add a GitHub action to sync data to the hub's S3 bucket
This action runs when a pull request is merged. It detects changes
to files (additions, updates, and deletes) and syncs those changes
to the hub's AWS S3 bucket (in this case, hosted on the Hubverse's
AWS account).  When first merged, the action will trigger a sync of
all existing, historical information to S3.

Going forward, only files that have changed will sync to S3.

Security is handled via a Github-specific OpenID Connect identity
provider in the Hubverse's AWS account, which grants temporary
write access to the S3 bucket from the main branch of this repo.
For more information about GitHub actions and AWS OpenID Connect:
https://docs.github.com/en/actions/deployment/security-hardening-your-deployments/configuring-openid-connect-in-amazon-web-services