explicitly use dplyr lag
This is future-proofing these statements in the case the functions here are ever moved to a package. 

The `lag()` function is in the {stats} package, which is attached in R by default. It is designed to work with `ts` objects. Importantly, it does _not_ do what you expect the {dplyr} incantation to do. 

If this function is ever moved into a context where {dplyr} is not explicitly attached via `library(dplyr)`, then this function will produce results that are not lagged.