Remove the as_of column from Hubverse target-data files
The read target data functions in hubData cannot read partitioned .csv files
that include paritioned columns in the file itself. That is a limitation
of Arrow (which powers hubData). While that limitation might be addressed
at a future date, for now, we'll have to rely the hive partitions to
return as_of information when reading the files.