Append new time series to existing file instead of partioning
This hub uses the .csv file format for target data, so we should not be
partioning it (per the Hubverse target-data specification).

Instead, complete versions of the new time series data should be appended
to the existing time-series.csv file.

This changeset also pass the location of target data to create_target_data
instead of assuming it will be "target-data" (in case we would want to do
something like pass in tempdir() when writing a test)